test(webapp): split triggerTask files to <=4 containerTests each
Follow the same reasoning one level further: each containerTest is a fresh
Postgres + Redis, so the per-file container cost scales with test count. Break
the two larger split files down so no file holds more than four cases, letting
the sequencer scatter container setups more evenly across shards.

- triggerTask.test.ts (3)          trigger + concurrencyKey coercion
- triggerTask.idempotency.test.ts  idempotency + queue resolution
- triggerTask.debounce.test.ts     retries + debounce validation
- triggerTask.mollifier.test.ts    mollifier call-site behaviour
- triggerTask.metadataCache.test.ts / triggerTask.residency.test.ts unchanged

23 cases total, unchanged. test-timings entries updated to match.
