+++ test-timings.json
@@ -76,10 +76,12 @@
   "apps/webapp/test/engine/batchPayloads.test.ts": 5018,
   "apps/webapp/test/engine/streamBatchItems.test.ts": 45000,
   "apps/webapp/test/engine/taskIdentifierRegistry.test.ts": 13152,
-  "apps/webapp/test/engine/triggerTask.debounce-mollifier.test.ts": 11002,
+  "apps/webapp/test/engine/triggerTask.debounce.test.ts": 5501,
+  "apps/webapp/test/engine/triggerTask.idempotency.test.ts": 5501,
   "apps/webapp/test/engine/triggerTask.metadataCache.test.ts": 5501,
+  "apps/webapp/test/engine/triggerTask.mollifier.test.ts": 5501,
   "apps/webapp/test/engine/triggerTask.residency.test.ts": 5501,
-  "apps/webapp/test/engine/triggerTask.test.ts": 9626,
+  "apps/webapp/test/engine/triggerTask.test.ts": 4125,
   "apps/webapp/test/presenters/mapRunToLiveFields.test.ts": 3,
   "apps/webapp/test/services/organizationAccessToken.test.ts": 9,
   "apps/webapp/test/services/personalAccessToken.test.ts": 8,
